Generator: Implement test generator
