+++ pisek/task_jobs/solution.py
@@ -32,22 +32,14 @@ from pisek.task_jobs.task_manager import TaskJobManager
 from pisek.task_jobs.program import RunResult, ProgramsJob
 from pisek.task_jobs.compile import Compile
 from pisek.task_jobs.generator.input_info import InputInfo
-from pisek.task_jobs.generator.base_classes import (
-    GenerateInput,
-    GeneratorTestDeterminism,
-)
-from pisek.task_jobs.generator.generator import (
-    generate_input,
-    generator_test_determinism,
-)
-from pisek.task_jobs.checker import CheckerJob
-from pisek.task_jobs.data import InputSmall, OutputSmall
+from pisek.task_jobs.data.data import OutputSmall
+from pisek.task_jobs.generator.manager import GenInputsMixin
 from pisek.task_jobs.tools import IsClean
 from pisek.task_jobs.solution_result import Verdict, SolutionResult
 from pisek.task_jobs.judge import judge_job, RunJudge, RunCMSJudge, RunBatchJudge
 
 
-class SolutionManager(TaskJobManager):
+class SolutionManager(TaskJobManager, GenInputsMixin):
     """Runs a solution and checks if it works as expected."""
 
     def __init__(self, solution_label: str, is_first: bool) -> None:
@@ -76,113 +68,51 @@ class SolutionManager(TaskJobManager):
 
         return jobs
 
-    def _input_info_jobs(self, input_info: InputInfo, subtask: int) -> list[Job]:
-        repeat = input_info.repeat * (self._env.inputs if input_info.seeded else 1)
-
-        jobs: list[Job] = []
-
-        random.seed(4)  # Reproducibility!
-        seeds = random.sample(range(0, 16**4), repeat)
-        for i, seed in enumerate(seeds):
-            inp_path = input_info.task_path(self._env, seed)
-            if inp_path in self._judges:
-                self.subtasks[-1].previous_jobs.append(self._judges[inp_path])
-                continue
-
-            input_jobs: list[Optional[Job]] = []
-            if self._is_first:
-                # test_seeded
-                input_jobs = list(
-                    self._generate_input_jobs(input_info, seed, subtask, i == 0)
-                )
-            output_jobs: list[Optional[Job]] = list(
-                self._solution_jobs(inp_path, subtask)
-            )
-
-            for j in input_jobs + output_jobs:
-                if j is not None:
-                    jobs.append(j)
-
-        return jobs
+    def _skip_input(self, input_info: InputInfo, seed: int, subtask: int) -> bool:
+        input_path = input_info.task_path(self._env, seed)
+        if input_path in self._judges:
+            self.subtasks[-1].previous_jobs.append(self._judges[input_path])
+            return True
+        else:
+            return False
 
     def _generate_input_jobs(
         self, input_info: InputInfo, seed: int, subtask: int, test_determinism: bool
-    ) -> tuple[
-        Optional[GenerateInput],
-        Optional[GeneratorTestDeterminism],
-        Optional[CheckerJob],
-        Optional[IsClean],
-        Optional[InputSmall],
-    ]:
-        if not input_info.is_generated:
-            return (None, None, None, None, None)
-
-        test_det: Optional[GeneratorTestDeterminism] = None
-        check_input: Optional[CheckerJob] = None
-        input_clean: Optional[IsClean] = None
-        input_small: Optional[InputSmall] = None
-
-        input_path = input_info.task_path(self._env, seed)
-        gen_inp = generate_input(self._env, self._env.config.in_gen, input_info, seed)
-
-        if input_info.seeded and test_determinism:
-            test_det = generator_test_determinism(
-                self._env, self._env.config.in_gen, input_info, seed
-            )
-            test_det.add_prerequisite(gen_inp)
-
-        if self._env.config.in_format == DataFormat.text:
-            input_clean = IsClean(self._env, input_path)
-            input_clean.add_prerequisite(gen_inp)
-
-        if self._env.config.limits.input_max_size != 0:
-            input_small = InputSmall(self._env, input_path)
-            input_small.add_prerequisite(gen_inp)
-
-        if self._env.config.checker is not None:
-            check_input = CheckerJob(
-                self._env,
-                self._env.config.checker,
-                input_path,
-                subtask,
-            )
-            check_input.add_prerequisite(gen_inp)
-
-        return (gen_inp, test_det, check_input, input_clean, input_small)
-
-    def _checker_jobs(self, input_path: TaskPath, subtask: int):
-        return None
+    ) -> list[Job]:
+        if not self._is_first:
+            return []
+        return super()._generate_input_jobs(input_info, seed, subtask, test_determinism)
 
     def _solution_jobs(
-        self, input_path: TaskPath, subtask: int
-    ) -> tuple[
-        "RunSolution", Optional["RunJudge"], Optional[IsClean], Optional[OutputSmall]
-    ]:
-        run_sol: RunSolution
-        run_judge: Optional[RunJudge] = None
-        out_clean: Optional[IsClean] = None
-        out_small: Optional[OutputSmall] = None
+        self, input_info: InputInfo, seed: int, subtask: int
+    ) -> list[Job]:
+        input_path = input_info.task_path(self._env, seed)
 
+        jobs: list[Job] = []
+
+        run_sol: RunSolution
+        run_judge: RunJudge
         if self._env.config.task_type == "batch":
             run_batch_sol, run_judge = self._create_batch_jobs(input_path, subtask)
             run_sol = run_batch_sol
-            self._judges[input_path] = run_judge
-            self.subtasks[-1].new_jobs.append(run_judge)
+            jobs += [run_batch_sol, run_judge]
 
             if self._env.config.out_format == DataFormat.text:
-                out_clean = IsClean(self._env, run_batch_sol.output)
+                jobs.append(out_clean := IsClean(self._env, run_batch_sol.output))
                 out_clean.add_prerequisite(run_judge)  # TODO: Conditional
             if self._env.config.limits.output_max_size != 0:
-                out_small = OutputSmall(self._env, run_batch_sol.output)
+                jobs.append(out_small := OutputSmall(self._env, run_batch_sol.output))
                 out_small.add_prerequisite(run_judge)
 
         elif self._env.config.task_type == "communication":
-            run_sol = self._create_communication_jobs(input_path)
-            self._judges[input_path] = run_sol
-            self.subtasks[-1].new_jobs.append(run_sol)
+            run_sol = run_judge = self._create_communication_jobs(input_path)
+            jobs.append(run_sol)
 
+        self._judges[input_path] = run_judge
+        self.subtasks[-1].new_jobs.append(run_judge)
         self.subtasks[-1].new_run_jobs.append(run_sol)
-        return (run_sol, run_judge, out_clean, out_small)
+
+        return jobs
 
     def _create_batch_jobs(
         self, inp: TaskPath, subtask: int
